pyproject.toml
[project]
name = "videofinal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "pillow>=12.0.0",
    "scipy>=1.15.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
